+++ CHANGELOG.md
@@ -1,12 +1,17 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+- Arguments in examples and tests.
+
+
 ## [7.3.2] - 2024-06-05
 
 ### Added
 - Set RPDOMap values by byte string.
 - Function to check if a configuration has been applied to the drive.
 - EtherCAT connection example script.
-- Arguments in examples and tests.
 
 ### Changed
 - The signature of the load_firmware method of EthercatNetwork is changed to add the boot_in_app argument.